// src/shared/types/api.ts
export type InitResponse = {
  type: 'init';
  postId: string;
  count: number;
  username: string;
  gameType: string;
};

export type IncrementResponse = {
  type: 'increment';
  postId: string;
  count: number;
};

export type DecrementResponse = {
  type: 'decrement';
  postId: string;
  count: number;
};